requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==23.2.0
    # via
    #   cattrs
    #   lsprotocol
cattrs==23.2.3
    # via
    #   lsprotocol
    #   python-lsp-ruff
docstring-to-markdown==0.15
    # via python-lsp-server
jedi==0.19.1
    # via python-lsp-server
lsprotocol==2023.0.1
    # via python-lsp-ruff
numpy==2.0.0
    # via -r requirements.in
parso==0.8.4
    # via jedi
pillow==10.4.0
    # via -r requirements.in
pluggy==1.5.0
    # via python-lsp-server
python-lsp-jsonrpc==1.1.2
    # via python-lsp-server
python-lsp-ruff==2.2.1
    # via -r requirements.in
python-lsp-server==1.11.0
    # via
    #   -r requirements.in
    #   python-lsp-ruff
ruff==0.5.0
    # via python-lsp-ruff
ujson==5.10.0
    # via
    #   python-lsp-jsonrpc
    #   python-lsp-server
